Refactor type definitions in types.h
Updated type definitions for u64 and s64 to use 'long long' for better compatibility.

+++ include/types.h
@@ -1,15 +1,15 @@
 #ifndef SHIFTOS_TYPES_H
 #define SHIFTOS_TYPES_H
 
-typedef unsigned char  u8;
+typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int   u32;
-typedef unsigned long  u64;
+typedef unsigned int u32;
+typedef unsigned long long u64;
 
-typedef signed char    s8;
-typedef signed short   s16;
-typedef signed int     s32;
-typedef signed long    s64;
+typedef signed char s8;
+typedef signed short s16;
+typedef signed int s32;
+typedef signed long long s64;
 
 typedef u64 size_t;
